requirements.txt
torch
torchvision
torchaudio
transformers
kaggle
wandb
tqdm
hydra-core
python-dotenv
datasets
sentencepiece
scikit-learn
pyspellchecker
lightgbm
nltk
spacy
gpl
optuna
textdistance
textstat